Kneeling Animation
Added kneeling and standing animation. They are now part of the diver
animator controller object. To get the diver to kneel set the bool
"kneel" to true. The diver will kneel. To have them stand up set the
bool "stand" to true and the bool "kneel" to false and the diver will
stand and then return to the idling animation.